+++ src/main/scala/org/podval/store/Path.scala
@@ -7,6 +7,9 @@ final case class Path(stores: Seq[Store]) derives CanEqual:
   def last: Store = stores.last
   def lastAs[T]: T = stores.last.asInstanceOf[T]
   def :+(store: Store): Path = Path(stores :+ store)
+  def tail: Path = Path(stores.tail)
+  def init: Path = Path(stores.init)
+  def parent: Path = init
   def structureNames: Seq[String] = stores.map(_.names.doFind(Language.English.toSpec).name)
   def toUrl: String = if stores.isEmpty then "/" else "/" + structureNames.mkString("/")
   override def toString: String = toUrl