+++ src/main/scala/org/podval/store/Stores.scala
@@ -11,7 +11,22 @@ trait Stores[+T <: Store] extends Store:
 
   def findByName(name: String): Option[T] = HasValues.find(stores, name)
 
-  // TODO add indexOf() and friends
+  def indexOf(store: Store): Int = stores.indexWhere(_ eq store)
+
+  final def next(store: Store): Option[T] =
+    val i: Int = indexOf(store)
+    Option.when(i >= 0 && i + 1 < stores.length)(stores(i + 1))
+
+  final def prev(store: Store): Option[T] =
+    val i: Int = indexOf(store)
+    Option.when(i > 0)(stores(i - 1))
+
+  final def distance(from: Store, to: Store): Int =
+    val fromI: Int = indexOf(from)
+    val toI: Int = indexOf(to)
+    require(fromI >= 0, s"Did not find $from in $this")
+    require(toI >= 0, s"Did not find $to in $this")
+    toI - fromI
 
   /*
   Successful `resolve()` returns a non-empty `Path`.